UBUNTU: [Packaging] linux-perf: Fix python perf library location
BugLink: https://bugs.launchpad.net/bugs/2103653

Now that we have a single unversioned linux-perf package per series,
there's only one python perf library and we no longer need a wrapper
to find the library that matches the running kernel. So put the library
in the correct location /usr/lib/python3/dist-packages/ so that Python
can find it on its own and remove the wrapper.

Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>
Acked-by: Kuba Pawlak <kuba.pawlak@canonical.com>
Acked-by: Kevin Becker <kevin.becker@canonical.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>